Beginning of Colors for spikes
I have created an enum for colors that can be collected. I've made it so that there's a bool that I named WOKE will be set to true when the color that unlocks it or any color afterward is unlocked, so that it will start acting. Not done yet, but will commit anyway.